Create docker_inspecz.py file to match upstream Update Dockerfile to match the new inspectz.py file name Create README.md to help users
